Allowing ABI to be stored in the blockchain
This change intends to strength the usage of NEP-3, by storing the ABI in the blockchain and allowing easy retrieval with RPC call "getcontractstate". This change is really simple and do not break compatibility with anything else, it just explores the description field, in order to store the ABI together with a raw description of the contract. For older deployed contracts, the ABI will be empty "{}", and newer contracts could use this field in the following way:
```
{
    "description": "Usual contract description",
    "ABI" : {
           // contract ABI
     }
}
```