SmartContract/Native: implement adjustable opcode pricing
This defines a new Policy contract setting, BaseExecFee which contains a price
of basic VM execution unit --- NOP instruction. It then redefines all opcode
prices as coefficients applied to that base fee.

The following invariants hold for this:
 * transaction fees are expressed in GAS
 * BaseExecFee is expressed in GAS
 * opcode price is a coefficient now
 * syscall price is expressed in GAS
 * ApplicationEngine.AddGas() also accepts GAS values, so it's a caller's
   responsibility to calculate that in whatever fashion he likes.

Caveats:
 * while coefficients are based on table from #1875 (which is built on
   instruction execution time relations for two complete VM implementations),
   some modifications were applied:
   - it's impossible for SYSCALL to have non-0 cost now (tests just hang)
   - all slot operations have the same price
 * block and consensus payloads are adjusted to use BaseExecFee, but probably
   ECDsaVerifyPrice is still a constant
 * it's not really tested other than unit tests

In general, this allows to define any execution cost in NOPs and then use
BaseExecFee to get the price. The same principle could be applied to storage
pricing based on FeePerByte (StoragePrice could just be `100 * FeePerByte`),
but that's a bit different topic.

Closes #1875.