Revert "Sort and limit size"
This reverts commit 2a6904784bd99e5493ab8adc9fe520e34b9a8b3a.

As it was said in https://github.com/neo-project/neo/pull/2909#issuecomment-1721564851,
we can't save only a part of on-chained conflicting signers. Instead, we
need to construct new blocks in such way so that there were no
conflicting signers limit overflow. It's not hard to implement, see the
further commits.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>