add a worst case test:
```
The regex pattern ^(a?)*a$ is known for catastrophic backtracking. The reason is the nested quantifier: (a?)*. This pattern can match zero or one occurrence of "a" any number of times. This leads to ambiguity when trying to match a string made up mostly of the character 'a'. The regex engine will try out all combinations of using the optional 'a' quantifier to see if there's a match.

For an input string made up of  n 'a' characters followed by another character (let's say 'b'), the regex engine would potentially have to explore 2^n possibilities.
```