Native: extend CryptoLib's verifyWithECDsa with hasher parameter
A port of
https://github.com/nspcc-dev/neo-go/pull/3425/commits/1e2b438b5580ca9ada7b452130b95a249733d31a.

This commit contains minor protocol extension needed for custom
Koblitz-based verification scripts (an alternative to
https://github.com/neo-project/neo/pull/3205).

Replace native CryptoLib's verifyWithECDsa `curve` parameter by
`curveHash` parameter which is a enum over supported pairs of named
curves and hash functions. NamedCurve enum mark as deprecated and
replaced by NamedCurveHash with compatible behaviour.

Even though this change is a compatible extension of the protocol, it
changes the genesis state due to parameter renaming (CryptoLib's
manifest is changed). But we're going to resync chain in 3.7 release
anyway, so it's not a big deal.

Also, we need to check mainnet and testnet compatibility in case if
anyone has ever called verifyWithECDsa with 24 or 25 `curve` value.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>