SmartContract: use executing contract state to check permissions
It's not correct to use an updated contract state got from native
Management to check for the allowed method call. We need to use
manifest from the currently executing context for that. It may be
critical for cases when executing contract is being updated firstly,
and after that it calls another contract. So we need an old (executing)
contract manifest for this check.

This change is moved under D hardfork to avoid state diff issues on
nodes update. Although it should be noted that it's hard to meet the
trigger criteria.

A port of https://github.com/nspcc-dev/neo-go/pull/3473. This bug was
discovered during the similar problem described in
https://github.com/nspcc-dev/neo-go/issues/3471 and fixed in
https://github.com/nspcc-dev/neo-go/pull/3472. I've checked all other
similar usages and the rest of them use proper contract state (executing
one, not the Management's one).

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>