P2P: fix `UnknownCapability`validation and maintainig (#3647)
* P2P: allow to serialize UnknownCapability

Otherwise it can't be properly passed through the P2P network which makes
it impossible to use UnknownCapabilities as an extensions in the existing
networks.

Follow https://github.com/nspcc-dev/neo-go/pull/3778.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>

* P2P: fix verification rules for NetworkAddressWithTime

Without this commit every `Addr` message containing UnknownCapability
will be considered as invalid. The desired behaviour is to include node
with UnknownCapability into the list of peers.

Follow the https://github.com/nspcc-dev/neo-go/pull/3778, should be a
part of https://github.com/neo-project/neo/pull/3639.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>

* P2P: don't strip UnknownCapabilities from node's version

Make the behaviour compatible with
https://github.com/nspcc-dev/neo-go/pull/3778.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>

---------

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>