Native: move MaxValidUntilBlockIncrement to native Policy
The problem is described in
https://github.com/nspcc-dev/neo-go/issues/3841. In short words,
the chain's state is dependent on MaxValidUntilBlockIncrement setting,
whereas this setting depends on the block time. Since block generation
time is a part of native Policy (#3622), MaxValidUntilBlockIncrement
should follow, we need the same value to be applied at all CNs
simultaneously.

Port https://github.com/nspcc-dev/neo-go/pull/3849.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>