Native: move `MaxValidUntilBlockIncrement` to native Policy (#3861)
* Native: move MaxValidUntilBlockIncrement to native Policy

The problem is described in
https://github.com/nspcc-dev/neo-go/issues/3841. In short words,
the chain's state is dependent on MaxValidUntilBlockIncrement setting,
whereas this setting depends on the block time. Since block generation
time is a part of native Policy (#3622), MaxValidUntilBlockIncrement
should follow, we need the same value to be applied at all CNs
simultaneously.

Port https://github.com/nspcc-dev/neo-go/pull/3849.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>

* Extensions: improve formatting

No functional changes.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>

* Native: set maximum constraint for MaxValidUntilBlockIncrement

It's set to be a day of 1-second blocks, that should be enough, ref.
https://github.com/neo-project/neo/pull/3861#discussion_r2027070961.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>

---------

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>
Co-authored-by: Shargon <shargon@gmail.com>
Co-authored-by: Will <201105916+Wi1l-B0t@users.noreply.github.com>
Co-authored-by: Christopher Schuchardt <cschuchardt88@gmail.com>
Co-authored-by: Jimmy <jinghui@wayne.edu>