Fixes sign command signatures and ambiguous calls bugs
Refactors the `sign message` command to:
- allow only signing only using the secp256r1 curve.
- output the algorithm used to create the payload
- output guide on how to verify the signed message

Renames the `sign` command to `sign transaction` to avoid ambiguous calls with the new `sign message` command.

Refactors the `OnCommand` method on the ConsoleServiceBase class so that parsing of arguments are differed to later in the execution. This is to solve an issue that occurs when ambiguous calls occur, one method might output errors in when parsing of arguments.